Complete antibody documents
Antibody documents now use the new document mechanism, and I deleted the old antibody characterization document component.
