+++ src/encoded/static/components/doc.js
@@ -123,7 +123,7 @@ var Document = module.exports.Document = React.createClass({
 
 
 // Document header component -- default
-var DocumentHeader = React.createClass({
+var DocumentHeader = module.exports.DocumentHeader = React.createClass({
     propTypes: {
         doc: React.PropTypes.object.isRequired // Document object to render
     },
@@ -139,26 +139,8 @@ var DocumentHeader = React.createClass({
     }
 });
 
-// Document header component -- Characterizations
-var CharacterizationHeader = React.createClass({
-    propTypes: {
-        doc: React.PropTypes.object.isRequired // Document object to render
-    },
-
-    render: function() {
-        var doc = this.props.doc;
-
-        return (
-            <div className="panel-header document-title sentence-case">
-                {doc.characterization_method}
-            </div>
-        );
-    }
-});
-
-
 // Document caption component -- default
-var DocumentCaption = React.createClass({
+var DocumentCaption = module.exports.DocumentCaption = React.createClass({
     propTypes: {
         doc: React.PropTypes.object.isRequired // Document object to render
     },
@@ -171,7 +153,7 @@ var DocumentCaption = React.createClass({
         }
 
         return (
-            <div className="document-intro document-meta-data key-value-left">
+            <div className="document-intro document-meta-data">
                 {excerpt || caption ?
                     <div data-test="caption">
                         <strong>{excerpt ? 'Description excerpt: ' : 'Description: '}</strong>
@@ -183,35 +165,8 @@ var DocumentCaption = React.createClass({
     }
 });
 
-// Document caption component -- Characterizations
-var CharacterizationCaption = React.createClass({
-    propTypes: {
-        doc: React.PropTypes.object.isRequired // Document object to render
-    },
-
-    render: function() {
-        var doc = this.props.doc;
-        var excerpt, caption = doc.caption;
-        if (caption && caption.length > EXCERPT_LENGTH) {
-            excerpt = globals.truncateString(caption, EXCERPT_LENGTH);
-        }
-
-        return (
-            <dl className="document-intro document-meta-data key-value-left">
-                {excerpt || caption ?
-                    <div data-test="caption">
-                        <strong>{excerpt ? 'Caption excerpt: ' : 'Caption: '}</strong>
-                        {excerpt ? <span>{excerpt}</span> : <span>{caption}</span>}
-                    </div>
-                : <em>No caption</em>}
-            </dl>
-        );
-    }
-});
-
-
 // Document preview component -- default
-var DocumentPreview = React.createClass({
+var DocumentPreview = module.exports.DocumentPreview = React.createClass({
     propTypes: {
         doc: React.PropTypes.object.isRequired // Document object to render
     },
@@ -227,7 +182,7 @@ var DocumentPreview = React.createClass({
 
 
 // Document file component -- default
-var DocumentFile = React.createClass({
+var DocumentFile = module.exports.DocumentFile = React.createClass({
     propTypes: {
         doc: React.PropTypes.object.isRequired, // Document object to render
         detailOpen: React.PropTypes.bool, // True if detail panel is visible
@@ -266,7 +221,7 @@ var DocumentFile = React.createClass({
 
 
 // Document detail component -- default
-var DocumentDetail = React.createClass({
+var DocumentDetail = module.exports.DocumentDetail = React.createClass({
     propTypes: {
         doc: React.PropTypes.object.isRequired, // Document object to render
         detailOpen: React.PropTypes.bool, // True if detail panel is visible
@@ -312,80 +267,20 @@ var DocumentDetail = React.createClass({
     }
 });
 
-// Document detail component -- default
-var CharacterizationDetail = React.createClass({
-    propTypes: {
-        doc: React.PropTypes.object.isRequired, // Document object to render
-        detailOpen: React.PropTypes.bool, // True if detail panel is visible
-        key: React.PropTypes.string // Unique key for identification
-    },
-
-    render: function() {
-        var doc = this.props.doc;
-        var keyClass = 'document-slider' + (this.props.detailOpen ? ' active' : '');
-        var excerpt = doc.description && doc.description.length > EXCERPT_LENGTH;
-
-        return (
-            <div className={keyClass}>
-                <dl className='key-value-doc' id={'panel' + this.props.key} aria-labeledby={'tab' + this.props.key} role="tabpanel">
-                    {excerpt ?
-                        <div data-test="caption">
-                            <dt>Caption</dt>
-                            <dd>{doc.caption}</dd>
-                        </div>
-                    : null}
-
-                    {doc.submitted_by && doc.submitted_by.title ?
-                        <div data-test="submitted-by">
-                            <dt>Submitted by</dt>
-                            <dd>{doc.submitted_by.title}</dd>
-                        </div>
-                    : null}
-
-                    <div data-test="lab">
-                        <dt>Lab</dt>
-                        <dd>{doc.lab.title}</dd>
-                    </div>
-
-                    {doc.award && doc.award.name ?
-                        <div data-test="award">
-                            <dt>Grant</dt>
-                            <dd>{doc.award.name}</dd>
-                        </div>
-                    : null}
-                </dl>
-            </div>
-        );
-    }
-});
-
-
 // Register document @types so they display in the standard document panel
 globals.panel_views.register(Document, 'Document');
-globals.panel_views.register(Document, 'BiosampleCharacterization');
-globals.panel_views.register(Document, 'DonorCharacterization');
 
 // Register document header rendering components
 globals.document_views.header.register(DocumentHeader, 'Document');
-globals.document_views.header.register(CharacterizationHeader, 'BiosampleCharacterization');
-globals.document_views.header.register(CharacterizationHeader, 'DonorCharacterization');
 
 // Register document caption rendering components
 globals.document_views.caption.register(DocumentCaption, 'Document');
-globals.document_views.caption.register(CharacterizationCaption, 'BiosampleCharacterization');
-globals.document_views.caption.register(CharacterizationCaption, 'DonorCharacterization');
 
 // Register document preview rendering components
 globals.document_views.preview.register(DocumentPreview, 'Document');
-globals.document_views.preview.register(DocumentPreview, 'BiosampleCharacterization');
-globals.document_views.preview.register(DocumentPreview, 'DonorCharacterization');
 
 // Register document file rendering components
 globals.document_views.file.register(DocumentFile, 'Document');
-globals.document_views.file.register(DocumentFile, 'BiosampleCharacterization');
-globals.document_views.file.register(DocumentFile, 'DonorCharacterization');
 
 // Register document detail rendering components
 globals.document_views.detail.register(DocumentDetail, 'Document');
-globals.document_views.detail.register(CharacterizationDetail, 'BiosampleCharacterization');
-globals.document_views.detail.register(CharacterizationDetail, 'DonorCharacterization');
